Create Selection sort
It is also a basic sorting technique like bubble sort. The basic idea is to find out the first minimum element in the array and place it to 1st and find the second minimum element in the array and place it to 2nd and go on until the array gets sorted.